Performs clustering of samples or features.

Performs clustering of samples or features.

clust(
  leaf_node = NULL,
  dat,
  clust_vec,
  clust_samps = TRUE,
  clust_feats = FALSE
)

Arguments

leaf_node
Integer value indicating terminal node id.

dat
Dataframe of the original dataset. Samples may be reordered.

clust_vec
Character vector of variable names to be applied clustering on. Can include class labels.

clust_samps
if TRUE, clusters samples within each leaf_node using Gower metric based on the values of the variables in clust_vec and returns the ordered dataframe for each leaf_node.

clust_feats
if TRUE clusters displayed features (passed through clust_vec) using the the Gower metric based on the values of all samples and returns the ordered features. When clust_samps = FALSE and clust_feats = FALSE, no clustering is performed.

Value

Dataframe of reordered original dataset when clust_samps == TRUE. Character vector of reordered features when clust_feats == TRUE.