Draws the heatmap.

Draws the heatmap to be placed below the decision tree.

draw_heat(
  dat,
  feat_names,
  disp_feats,
  class_cols,
  panel_space,
  feat_types = NULL,
  trans_type = "normalize",
  cont_cols = ggplot2::scale_fill_viridis_c(),
  cate_cols = ggplot2::scale_fill_viridis_d(option = "D", begin = 0.3, end = 0.9),
  clust_feats = TRUE,
  class_space = 0.03,
  class_pos = "top",
  class_lab_disp = NULL
)

Arguments

dat
Dataframe with samples from original dataset ordered according to the clustering within each leaf node.

feat_names
Character vector specifying the feature names in dat.

disp_feats
Character vector specifying features to be displayed.

class_cols
Vector of RGBs for the class colors, defaults to a colorblind friendly palette.

panel_space
Spacing between facets relative to viewport, recommended to range from 0.001 to 0.01.

feat_types
Named vector indicating the type of each features, e.g., c(sex = 'factor', age = 'numeric'). If feature types are not supplied, infer from column type.

trans_type
Character string specifying transformation type, can be 'scale' or 'normalize'.

cont_cols
Function determine color scale for continuous variable, defaults to viridis option D.

cate_cols
Function determine color scale for nominal categorical variable, defaults to viridis option D.

clust_feats
Logical. If TRUE, performs cluster on the features.

class_space
Numeric value indicating spacing between the class label and the rest of the features

class_pos
Character string specifying the position of the class label on heatmap, can be 'top', 'bottom' or 'none'.

class_lab_disp
Character string for displaying the label of class label if it differs from class_lab.

Value

A ggplot2 grob object of the heatmap.