Draws and aligns decision tree and heatmap.

Draws and aligns decision tree and heatmap.

heat_tree(
  data,
  target_lab,
  data_test = NULL,
  task = c("classification", "regression"),
  target_cols = NULL,
  label_map = NULL,
  custom_layout = NULL,
  clust_samps = TRUE,
  clust_target = TRUE,
  show_all_feats = FALSE,
  p_thres = 0.05,
  lev_fac = 1.3,
  heat_rel_height = 0.2,
  tree_space_top = 0.05,
  tree_space_bottom = 0.05,
  par_node_vars = list(label.size = 0, label.padding = ggplot2::unit(0.15, "lines"),
    line_list = list(ggplot2::aes(label = splitvar)), line_gpar = list(list(size = 9)),
    ids = "inner"),
  terminal_vars = list(label.padding = ggplot2::unit(0.25, "lines"), size = 3, col =
    "white"),
  edge_vars = list(color = "grey70", size = 0.5),
  edge_text_vars = list(color = "grey30", size = 3, mapping = ggplot2::aes(label =
    paste(breaks_label, "*NA"))),
  feat_types = NULL,
  trans_type = c("normalize", "scale", "none"),
  cont_cols = ggplot2::scale_fill_viridis_c(),
  cate_cols = ggplot2::scale_fill_viridis_d(),
  clust_feats = TRUE,
  target_space = 0.05,
  panel_space = 0.001,
  target_pos = "top",
  target_lab_disp = target_lab,
  ...
)

Arguments

data
Tidy dataset.

target_lab
Name of the column in data that contains target/label information.

data_test
Tidy test dataset. If NULL, heatmap displays (training) `data`.

task
Character string indicating the type of problem, either 'classification' (categorical outcome) or 'regression' (continuous outcome).

target_cols
Function determine color scale for target, defaults to viridis option B.

label_map
Named vector of the meaning of the target values, e.g., c(`0` = 'Edible', `1` = 'Poisonous').

custom_layout
Dataframe with 3 columns: id, x and y for manually input custom layout.

clust_samps
Logical. If TRUE, hierarhical clustering would be performed among samples within each leaf node.

clust_target
Logical. If TRUE, target/label is included in hierarchical clustering of samples within each leaf node and might yield a more interpretable heatmap.

show_all_feats
Logical. If TRUE, show all features regarless p_thres.

p_thres
Numeric value indicating the p-value threshold of feature importance. Feature with p-values computed from the decision tree below this value will be displayed on the heatmap.

lev_fac
Relative weight of child node positions according to their levels, commonly ranges from 1 to 1.5. 1 for parent node perfectly in the middle of child nodes.

heat_rel_height
Relative height of heatmap compared to whole figure (with tree).

tree_space_top
Numeric value to pass to expand for top margin of tree.

tree_space_bottom
Numeric value to pass to expand for bottom margin of tree.

par_node_vars
Named list containing arguments to be passed to the `geom_node_label()` call for non-terminal nodes.

terminal_vars
Named list containing arguments to be passed to the `geom_node_label()` call for terminal nodes.

edge_vars
Named list containing arguments to be passed to the `geom_edge()` call for tree edges.

edge_text_vars
Named list containing arguments to be passed to the `geom_edge_label()` call for tree edge annotations.

feat_types
Named vector indicating the type of each features, e.g., c(sex = 'factor', age = 'numeric'). If feature types are not supplied, infer from column type.

trans_type
Character string of 'normalize', 'scale' or 'none'. If 'scale', subtract the mean and divide by the standard deviation. If 'normalize', i.e., max-min normalize, subtract the min and divide by the max. If 'none', no transformation is applied. More information on what transformation to choose can be acquired here: https://cran.rstudio.com/package=heatmaply/vignettes/heatmaply.html#data-transformation-scaling-normalize-and-percentize

cont_cols
Function determine color scale for continuous variable, defaults to viridis option D.

cate_cols
Function determine color scale for nominal categorical variable, defaults to viridis option D.

clust_feats
Logical. If TRUE, performs cluster on the features.

target_space
Numeric value indicating spacing between the target label and the rest of the features

panel_space
Spacing between facets relative to viewport, recommended to range from 0.001 to 0.01.

target_pos
Character string specifying the position of the target label on heatmap, can be 'top', 'bottom' or 'none'.

target_lab_disp
Character string for displaying the label of target label if not supplied, differs from target_lab.

Value

A gtable/grob object of the decision tree (top) and heatmap (bottom).

Examples

heat_tree(iris, 'Species')

#> TableGrob (14 x 15) "layout": 28 grobs
#>     z         cells       name                                    grob
#> 1   0 ( 2-14, 1-15) background zeroGrob[plot.background..zeroGrob.121]
#> 2   1 ( 9- 9, 5- 5)  panel-1-1                 gTree[panel-1.gTree.36]
#> 3   1 ( 9- 9, 7- 7)  panel-2-1                 gTree[panel-2.gTree.46]
#> 4   1 ( 9- 9, 9- 9)  panel-3-1                 gTree[panel-3.gTree.56]
#> 5   1 ( 9- 9,11-11)  panel-4-1                 gTree[panel-4.gTree.66]
#> 6   3 ( 7- 7, 5- 5)   axis-t-1                          zeroGrob[NULL]
#> 7   3 ( 7- 7, 7- 7)   axis-t-2                          zeroGrob[NULL]
#> 8   3 ( 7- 7, 9- 9)   axis-t-3                          zeroGrob[NULL]
#> 9   3 ( 7- 7,11-11)   axis-t-4                          zeroGrob[NULL]
#> 10  3 (10-10, 5- 5)   axis-b-1      absoluteGrob[GRID.absoluteGrob.67]
#> 11  3 (10-10, 7- 7)   axis-b-2      absoluteGrob[GRID.absoluteGrob.68]
#> 12  3 (10-10, 9- 9)   axis-b-3      absoluteGrob[GRID.absoluteGrob.69]
#> 13  3 (10-10,11-11)   axis-b-4      absoluteGrob[GRID.absoluteGrob.70]
#> 14  3 ( 9- 9, 4- 4)   axis-l-1      absoluteGrob[GRID.absoluteGrob.74]
#> 15  3 ( 9- 9,12-12)   axis-r-1                          zeroGrob[NULL]
#> 16  2 ( 8- 8, 5- 5)  strip-t-1 zeroGrob[strip.text.x.top..zeroGrob.75]
#> 17  2 ( 8- 8, 7- 7)  strip-t-2 zeroGrob[strip.text.x.top..zeroGrob.76]
#> 18  2 ( 8- 8, 9- 9)  strip-t-3 zeroGrob[strip.text.x.top..zeroGrob.77]
#> 19  2 ( 8- 8,11-11)  strip-t-4 zeroGrob[strip.text.x.top..zeroGrob.78]
#> 20  4 ( 6- 6, 5-11)     xlab-t                          zeroGrob[NULL]
#> 21  5 (11-11, 5-11)     xlab-b                          zeroGrob[NULL]
#> 22  6 ( 9- 9, 3- 3)     ylab-l                          zeroGrob[NULL]
#> 23  7 ( 9- 9,13-13)     ylab-r                          zeroGrob[NULL]
#> 24  8 ( 5- 5, 5-11)   subtitle   zeroGrob[plot.subtitle..zeroGrob.118]
#> 25  9 ( 4- 4, 5-11)      title      zeroGrob[plot.title..zeroGrob.117]
#> 26 10 (12-12, 5-11)    caption    zeroGrob[plot.caption..zeroGrob.120]
#> 27 11 ( 3- 3, 2- 2)        tag        zeroGrob[plot.tag..zeroGrob.119]
#> 28 12 ( 1- 1, 5-11)     layout                          gtable[layout]

heat_tree(
  data = galaxy[1:100, ],
  target_lab = 'target',
  task = 'regression',
  terminal_vars = NULL,
  tree_space_bottom = 0)

#> Warning: binary variable(s) 3 treated as interval scaled

#> TableGrob (14 x 21) "layout": 40 grobs
#>     z         cells       name                                     grob
#> 1   0 ( 2-14, 1-21) background  zeroGrob[plot.background..zeroGrob.306]
#> 2   1 ( 9- 9, 5- 5)  panel-1-1                 gTree[panel-1.gTree.197]
#> 3   1 ( 9- 9, 7- 7)  panel-2-1                 gTree[panel-2.gTree.207]
#> 4   1 ( 9- 9, 9- 9)  panel-3-1                 gTree[panel-3.gTree.217]
#> 5   1 ( 9- 9,11-11)  panel-4-1                 gTree[panel-4.gTree.227]
#> 6   1 ( 9- 9,13-13)  panel-5-1                 gTree[panel-5.gTree.237]
#> 7   1 ( 9- 9,15-15)  panel-6-1                 gTree[panel-6.gTree.247]
#> 8   1 ( 9- 9,17-17)  panel-7-1                 gTree[panel-7.gTree.257]
#> 9   3 ( 7- 7, 5- 5)   axis-t-1                           zeroGrob[NULL]
#> 10  3 ( 7- 7, 7- 7)   axis-t-2                           zeroGrob[NULL]
#> 11  3 ( 7- 7, 9- 9)   axis-t-3                           zeroGrob[NULL]
#> 12  3 ( 7- 7,11-11)   axis-t-4                           zeroGrob[NULL]
#> 13  3 ( 7- 7,13-13)   axis-t-5                           zeroGrob[NULL]
#> 14  3 ( 7- 7,15-15)   axis-t-6                           zeroGrob[NULL]
#> 15  3 ( 7- 7,17-17)   axis-t-7                           zeroGrob[NULL]
#> 16  3 (10-10, 5- 5)   axis-b-1      absoluteGrob[GRID.absoluteGrob.258]
#> 17  3 (10-10, 7- 7)   axis-b-2      absoluteGrob[GRID.absoluteGrob.259]
#> 18  3 (10-10, 9- 9)   axis-b-3      absoluteGrob[GRID.absoluteGrob.260]
#> 19  3 (10-10,11-11)   axis-b-4      absoluteGrob[GRID.absoluteGrob.261]
#> 20  3 (10-10,13-13)   axis-b-5      absoluteGrob[GRID.absoluteGrob.262]
#> 21  3 (10-10,15-15)   axis-b-6      absoluteGrob[GRID.absoluteGrob.263]
#> 22  3 (10-10,17-17)   axis-b-7      absoluteGrob[GRID.absoluteGrob.264]
#> 23  3 ( 9- 9, 4- 4)   axis-l-1      absoluteGrob[GRID.absoluteGrob.267]
#> 24  3 ( 9- 9,18-18)   axis-r-1                           zeroGrob[NULL]
#> 25  2 ( 8- 8, 5- 5)  strip-t-1 zeroGrob[strip.text.x.top..zeroGrob.268]
#> 26  2 ( 8- 8, 7- 7)  strip-t-2 zeroGrob[strip.text.x.top..zeroGrob.269]
#> 27  2 ( 8- 8, 9- 9)  strip-t-3 zeroGrob[strip.text.x.top..zeroGrob.270]
#> 28  2 ( 8- 8,11-11)  strip-t-4 zeroGrob[strip.text.x.top..zeroGrob.271]
#> 29  2 ( 8- 8,13-13)  strip-t-5 zeroGrob[strip.text.x.top..zeroGrob.272]
#> 30  2 ( 8- 8,15-15)  strip-t-6 zeroGrob[strip.text.x.top..zeroGrob.273]
#> 31  2 ( 8- 8,17-17)  strip-t-7 zeroGrob[strip.text.x.top..zeroGrob.274]
#> 32  4 ( 6- 6, 5-17)     xlab-t                           zeroGrob[NULL]
#> 33  5 (11-11, 5-17)     xlab-b                           zeroGrob[NULL]
#> 34  6 ( 9- 9, 3- 3)     ylab-l                           zeroGrob[NULL]
#> 35  7 ( 9- 9,19-19)     ylab-r                           zeroGrob[NULL]
#> 36  8 ( 5- 5, 5-17)   subtitle    zeroGrob[plot.subtitle..zeroGrob.303]
#> 37  9 ( 4- 4, 5-17)      title       zeroGrob[plot.title..zeroGrob.302]
#> 38 10 (12-12, 5-17)    caption     zeroGrob[plot.caption..zeroGrob.305]
#> 39 11 ( 3- 3, 2- 2)        tag         zeroGrob[plot.tag..zeroGrob.304]
#> 40 12 ( 1- 1, 5-17)     layout                           gtable[layout]